test all endpoints & fix bugs
